Rename safeStringify.ts to stringify.ts, lint
